Put security council verify as a struct method not free function
