+++ crates/sovereign-sdk/rollup-interface/Cargo.toml
@@ -37,6 +37,8 @@ tracing = { workspace = true, optional = true }
 sov-keys = { path = "../module-system/sov-keys" }
 
 [dev-dependencies]
+alloy-signer = { workspace = true }
+alloy-signer-local = { workspace = true }
 serde_json = { workspace = true }
 
 [features]
